Change DNS resolution method to avoid unnecessary reverse lookup
Dns.GetHostEntry() method does a reverse lookup on IP to try to obtain the hostname.
Once the hostname obtained it then query again the DNS to get all possible IPs. Hence there is 3 DNS lookup.
It is not necessary since the hostname is not used at all.

Hence we move to Dns.GetHostAddresses() method:
"The GetHostAddresses method has different behavior with respect to IP literals.
If step #1 above succeeds (it successfully parses as an IP address), that address is immediately returned as the result.
There is no attempt at a reverse lookup."

Please refer to Microsoft official documention:
https://msdn.microsoft.com/en-us/library/ms143998(v=vs.110).aspx
https://msdn.microsoft.com/en-us/library/system.net.dns.gethostaddresses(v=vs.110).aspx

Change-Id: I11f91421e5e959f44904f80a47b483bef22fb5f6