Gather loggers implem and make them public
The base idea is that I want to use a ConsoleLogger in my flaky tests, to ease
fixing them.
There's already a ConsoleLogger, but it's a private inner class. So I make it
public and I put it in its own file, in a main project.
It's then cleaner to also put the DevNullLogger along this implementation.

Change-Id: I0db0d87d60a2f06a6f5bbfc0a944c052ca6d6370