Fix flaky TestCluster.TestAllNodesDead
It sometimes failed when it was asserting that Node.FetchMetadata was called
twice. It's because the call to Cluster.RequireNewRoutingTable sometimes didn't
end up triggering a reload of the metadata.
This case could happen if it took less than 1s between the beginning of the test
of the call to RequireNewRoutingTable.

This patch ensure this test will always pass.
The impact on the prod code in negligible (and if clients increases the
MinimumTimeBetweenRefreshMetadata config by 1, then it's completely iso)

Change-Id: Id06867f8f68b053dcccfba0c2d261bc063e21146