Support for Kafka 0.10
- Support 0.10.1 binary protocol
- Still support 0.8.2 protocol

This is activated via a flag in the Configuration class. No support for
intermediary protocol, either you run in 0.8.2 compatibility mode and only
v0 of APIs will be used (except for consumer group support, which requires
kafka 0.10), or you run in 0.10+ compatibility mode and latest version of
APIs will be used (with support of consumer groups). Obviously 0.10
compatibillity only works with 0.10+ brokers. 0.8.2 compatibility works fine
with any kafka version >= 0.8.2.

Support for various versions of the APIs is not hardwired, that means the
non public classes do not really enforce consistency between different
APIs.

Change-Id: Ib37ad5ae7ae55fbb5ee3668585ff2dc55c005eb7