In case of deserialization error, consume from latest.
In case where we are trying to deserialize a message set, and inside of this
message set there is a badly formated message (from the protocol point of view),
it is really unclear what we should do.

The previous behaviour was to retry. Since this can lead to infinite loop, with
the new behaviour, if the configuration "ErrorStrategy" is set to "Discard",
then after such a problem the driver will start consuming from latest (thus not
requesting the faulty message anymore).

Change-Id: I022a590389acae52ab288140f89aa49dff7df412
JIRA: WBSC-3077