Fix a bug that prevented refreshing metadata on a dead node event
When detecting that a node was dead, a new routing table was created
that was based on the old one, just filtering the dead node partitions.

The problem was that at this stage the "LastRefresh" variable was set
to "Now", and the subsequent attempt to referesh metadata failed because
it was less than 42 secondes since last metadata refreshed.