In case metadata request fail, repost it at the cluster level.
This change is here to balance out the change made in the previous commit,
in order to not completely discard metadata request.

Since the tests are running with a dedicated scheduler that run the agent
synchronously on the same thread, always failing on a node would mean
infinite loop from the cluster; so we modify the concerned tests.