Fix a bug where the driver was blocked fetching metadata from a dead node.
The scenario of the bug was the following:
- 1 node is dead (let's call it node_1)
- trying to produce on partition that have node_1 as replica (but not
  necessarily leader) will fail. When node_1 is not a leader, this
  failure is a timeout on the produce from the (leader) broker side.
- when receiving such error message, the driver will try to refresh its
  metadata. At this point the driver is not necessarily aware that node_1
  is dead (it requires 3 successive connection failure (e.g. socket
  connection error) to declare the node as dead. 1 connection failure can
  take 20 sec at times).
- with bad luck (it is random), the driver is deciding to fetch its metadata
  from node_1.
  The producer is blocked until it gets metadata (we will call this problem1)
- as long as node_1 is not considered dead, if a request for metadata fails
  then it is reposted on the same node (we will call this problem2)
- during this time, messages received are not routed to any particular node.
  The driver overflow and/or the messages expire.

Problem1 is the biggest problem. It means that if something goes wrong anywhere
with the metadata (which happens in the case of problem2) then the whole driver
is blocked. We solve problem1 by adding a timeout on fetching metadata, and
we accept that if we do not get metadata in a reasonable time then we move
forward with stale metadata (which is better than blocking).

Problem2 is the specific problem that happened in the scenario. We solve it
by not reposting metadata request in case of connection error, failing faster.

In a future commit, the third thing that we want to do, to balance out the
previous changes, is to repost at the cluster level (so not the node level)
a metadata request in case the previous one fails. It should (at least
eventually) be routed to a node that is not node_1.