Fix bug where we query metadata from a broker that has no topology
We are not 100% sure if this is the case, but it seems like sometimes
when a node gets down and back up, and we are querying it at this exact
moment to get metadata before itself was able to fetch the topology of the
cluster, then it will answer us with an empty topology.

If that is the case we should not be satisfied (as this cause lot of
problem) but consider the request as failed and try again.