Add a dynamic seeds list.
The previous behaviour was to give the driver at initialisation a list
of seed broker to start discovering the topology of the cluster.

The problem is that this static list of brokers is used again in case
we end up during the life of the driver with a topology containing
0 broker.

At this point, the brokers of the initial list might have all been retired.
It makes more sense to have a way to get this external list of server
dynamically.