Change the error message when broker remotely close an idle connection.
Historically speaking when a broker was remotely closing an idle
connection we would treat it as a transport error.
We initally logged it as a warning.

In 52c777d9bceca5d3aa9d2cf160159c7489713f6a, we decided to start
logging it as an info.

This commit goes a bit further, by not treating it as an error at
all and just displaying a less threatning info log