Fix a bug where the state of the cluster is corrupted.
The bug was introduced in 2a56234a1ae56820da23f0e76a665addc1a0036e,
where we changed the behaviour of processing of node events to be
asynchrone.

We forgot an await, and inside the processing of a dead node event
we are modifying the dictionnary _nodes. Since we are not awaiting,
the execution continues and the agent is free to handle another
task (that may involve reading _nodes). This can lead to concurent
acces to the dictionnary and corrupt the state.