Implement production of records with headers
Expose the headers of the new records format to the public api.

RecordHeader is deleted and its methods are moved into Record.
We serialize/deserialize a public KafkaRecordHeader which is a struct
holding only a key and a value.

The value is always a byte array, avoiding the burden of managing the
serializers.

Change-Id: Idc8ef1e739399a4881a495c941cc29a30134c24e