[Kafka] Introduce partition selector with message key
Currently, there is only one partition selection mode, which is
the round robin one.

There's a need to introduce a new selection strategy depending on
the keys of messages to determine the partition the messages should
go to. It would allow to compact the same messages and reduce heavily
the number of messages.

This new partition selection strategy guarantees that messages with
the same key will be sent to the same partition, given the same
number of available partitions. It is not using a consistent hashing
though, which means that if new nodes are introduced or old ones are
removed from the cluster, new messages with the same key as old
messages will not be sent to the same partition.

JIRA: WBSC-3909

Change-Id: Ie75b40123fd90550aecafd9c33d2b82952470469