Handle brokers sending partial messages in FetchResponse
From the official documentation: "As an optimization the server is allowed
to return a partial message at the end of the message set. Clients should
handle this case.".

We then discard the message, the read offset not being updated for this message,
meaning we will get it in full in the next FetchRequest.

Change-Id: I0dae2e93641c7a48544d5a3b02287bd10319ab58