Introduce retry delay when fetching partitions in HandleStart fails
Currently, when trying to start consuming a topic, if the first fetch of
the available partitions fails (because the topic does not exist for example),
the driver will immediately retry. This can translate to hundreds of requests
per second, which is useless.

As this loop in HandleStart is already blocking indefinitely the ProcessMessage()
loop of the driver, introducing a retry delay does not change the behavior.

Change-Id: Ice89ab8e4cb1a41ba25fb2bec49bd1be66001a8b